+++ packages/ai-chat/src/chat/components-legacy/MessageTypeComponent.tsx
@@ -41,12 +41,12 @@ import { CarouselItemComponent } from "../components/messages/CarouselItemCompon
 import { ConversationalSearch } from "./responseTypes/conversationalSearch/ConversationalSearch";
 import UserDefinedResponse from "../components/UserDefinedResponse";
 import CustomFooterSlot from "./responseTypes/custom/CustomFooterSlot";
-import { DatePickerComponent } from "./responseTypes/datePicker/DatePickerComponent";
+import { DatePickerComponent } from "../components/responseTypes/datePicker/DatePickerComponent";
 import InlineError from "../components/util/InlineError";
 import { GridItemComponent } from "./responseTypes/grid/GridItemComponent";
 import { IFrameMessage } from "./responseTypes/iframe/IFrameMessage";
 import { Image } from "./responseTypes/image/Image";
-import { OptionComponent } from "./responseTypes/options/OptionComponent";
+import { OptionComponent } from "../components/responseTypes/options/OptionComponent";
 import { MarkdownWithErrorHandling } from "../components/util/MarkdownWithErrorHandling";
 import { VideoComponent } from "../components/messages/VideoComponent";
 import { useSelector } from "../hooks/useSelector";
