+++ packages/ai-chat/src/chat/components-legacy/responseTypes/_imports.scss
@@ -9,8 +9,6 @@
 @use "previewCard/imports" as responseType-previewCard;
 @use "carousel/imports" as responseType-carousel;
 @use "conversationalSearch/imports" as responseType-conversationalSearch;
-@use "datePicker/imports" as responseType-datePicker;
 @use "grid/imports" as responseType-grid;
 @use "iframe/imports" as responseType-iframe;
 @use "image/imports" as responseType-image;
-@use "options/imports" as responseType-options;
