+++ packages/ai-chat/src/chat/components/responseTypes/_imports.scss
@@ -1,8 +1,9 @@
 /* 
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *  
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
  */
 
-@use "DatePicker";
+@use "datePicker/imports" as responseType-datePicker;
+@use "options/imports" as responseType-options;
