+++ packages/ai-chat/src/chat/components/responseTypes/datePicker/DatePickerComponent.tsx
@@ -1,5 +1,5 @@
 /*
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
@@ -25,7 +25,7 @@ import {
   clearDynamicStyles,
 } from "../../../utils/cspStyleUtils";
 
-import { ScrollElementIntoViewFunction } from "../../MessagesComponent";
+import { ScrollElementIntoViewFunction } from "../../../components-legacy/MessagesComponent";
 import { useOnMount } from "../../../hooks/useOnMount";
 import { useServiceManager } from "../../../hooks/useServiceManager";
 import { AppState } from "../../../../types/state/AppState";
