+++ packages/ai-chat/src/chat/components/responseTypes/datePicker/_imports.scss
@@ -1,9 +1,8 @@
 /* 
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *  
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
  */
 
-@use "OptionComponent";
-@use "SelectComponent";
+@use "DatePicker";
