+++ packages/ai-chat/src/chat/components/responseTypes/options/OptionComponent.tsx
