+++ packages/ai-chat/src/chat/components/responseTypes/options/SelectComponent.tsx
