+++ packages/ai-chat/src/chat/components/responseTypes/options/_imports.scss
@@ -1,0 +1,9 @@
+/* 
+ *  Copyright IBM Corp. 2025, 2026
+ *  
+ *  This source code is licensed under the Apache-2.0 license found in the
+ *  LICENSE file in the root directory of this source tree.
+ */
+
+@use "OptionComponent";
+@use "SelectComponent";
